feat: refactor shipping components, models, and services into separate ‘shipping’ directory
