+++ src/app/app-routing.module.ts
@@ -1,14 +1,14 @@
 import { AdminAuthGuard } from './admin/services/admin-auth-guard.service';
-import { MyOrdersComponent } from './my-orders/my-orders.component';
+import { MyOrdersComponent } from './shopping/components/my-orders/my-orders.component';
 import { AdminProductsComponent } from './admin/components/admin-products/admin-products.component';
 import { LoginComponent } from './login/login.component';
-import { OrderSuccessComponent } from './order-success/order-success.component';
-import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
-import { ProductsComponent } from './products/products.component';
+import { OrderSuccessComponent } from './shopping/components/order-success/order-success.component';
+import { ShoppingCartComponent } from './shopping/components/shopping-cart/shopping-cart.component';
+import { ProductsComponent } from './shopping/components/products/products.component';
 import { HomeComponent } from './home/home.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { CheckOutComponent } from './check-out/check-out.component';
+import { CheckOutComponent } from './shopping/components/check-out/check-out.component';
 import { AdminOrdersComponent } from './admin/components/admin-orders/admin-orders.component';
 import { AuthGuard } from 'shared/services/auth-guard.service';
 import { ProductFormComponent } from './admin/components/product-form/product-form.component';
