+++ src/app/shopping/components/products/products.component.ts
