+++ src/app/shopping/components/shopping-cart-summary/shopping-cart-summary.component.ts
