+++ src/app/shopping/components/shopping-cart/shopping-cart.component.ts
